* commit '31e6948d897c8d6e00326d07f20d8dccccaecf4a':
  TransportMediator: handle playback position.

+++ v4/java/android/support/v4/media/TransportController.java
@@ -52,18 +52,18 @@ public abstract class TransportController {
     /**
      * Retrieve the total duration of the media stream, in milliseconds.
      */
-    public abstract int getDuration();
+    public abstract long getDuration();
 
     /**
      * Retrieve the current playback location in the media stream, in milliseconds.
      */
-    public abstract int getCurrentPosition();
+    public abstract long getCurrentPosition();
 
     /**
      * Move to a new location in the media stream.
      * @param pos Position to move to, in milliseconds.
      */
-    public abstract void seekTo(int pos);
+    public abstract void seekTo(long pos);
 
     /**
      * Return whether the player is currently playing its stream.
